sam-app/apis/requirements.txt
requests
asyncpg==0.29.0
python-dotenv==1.0.1
pydantic==2.8.2
pydantic_core==2.20.1
google-auth==2.30.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.135.0
pyjwt
bcrypt